requirements.txt
llama-cpp-python 
fastapi 
uvicorn 
streamlit 
requests
huggingface-hub